fix some more 64bit warnings/issues
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3795 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/cheat.c
@@ -2024,12 +2024,12 @@ static void RebuildStringTables(void)
 					menuStrings.mainStringLength,
 					menuStrings.subStringLength,
 
-					(int)menuStrings.mainList,
-					(int)menuStrings.subList,
-					(int)menuStrings.flagList,
-					(int)menuStrings.mainStrings,
-					(int)menuStrings.subStrings,
-					(int)menuStrings.buf);
+					(size_t)menuStrings.mainList,
+					(size_t)menuStrings.subList,
+					(size_t)menuStrings.flagList,
+					(size_t)menuStrings.mainStrings,
+					(size_t)menuStrings.subStrings,
+					(size_t)menuStrings.buf);
 
 		exit(1);
 	}
@@ -7896,7 +7896,7 @@ static void RestoreRegionBackup(SearchRegion * region)
 static UINT8 DefaultEnableRegion(SearchRegion * region, SearchInfo * info)
 {
 	mem_write_handler	handler = region->writeHandler->handler;
-	UINT32				handlerAddress = (UINT32)handler;
+	size_t				handlerAddress = (size_t)handler;
 
 	switch(info->searchSpeed)
 	{
@@ -7958,8 +7958,8 @@ static UINT8 DefaultEnableRegion(SearchRegion * region, SearchInfo * info)
 			return 0;
 
 		case kSearchSpeed_Medium:
-			if(	(handlerAddress >= ((UINT32)MWA_BANK1)) &&
-				(handlerAddress <= ((UINT32)MWA_BANK24)))
+			if(	(handlerAddress >= ((size_t)MWA_BANK1)) &&
+				(handlerAddress <= ((size_t)MWA_BANK24)))
 				return 1;
 
 			if(handler == MWA_RAM)
@@ -8000,24 +8000,24 @@ static void SetSearchRegionDefaultName(SearchRegion * region)
 			if(region->writeHandler)
 			{
 				mem_write_handler	handler = region->writeHandler->handler;
-				UINT32				handlerAddress = (UINT32)handler;
+				size_t				handlerAddress = (size_t)handler;
 
-				if(	(handlerAddress >= ((UINT32)MWA_BANK1)) &&
-					(handlerAddress <= ((UINT32)MWA_BANK24)))
+				if(	(handlerAddress >= ((size_t)MWA_BANK1)) &&
+					(handlerAddress <= ((size_t)MWA_BANK24)))
 				{
-					sprintf(desc, "BANK%.2d", (handlerAddress - ((UINT32)MWA_BANK1)) + 1);
+					sprintf(desc, "BANK%.2d", (handlerAddress - ((size_t)MWA_BANK1)) + 1);
 				}
 				else
-					if(handlerAddress == (UINT32)MWA_NOP)
+					if(handlerAddress == (size_t)MWA_NOP)
 						strcpy(desc, "NOP   ");
 					else
-					if(handlerAddress == (UINT32)MWA_RAM)
+					if(handlerAddress == (size_t)MWA_RAM)
 						strcpy(desc, "RAM   ");
 					else
-					if(handlerAddress == (UINT32)MWA_ROM)
+					if(handlerAddress == (size_t)MWA_ROM)
 						strcpy(desc, "ROM   ");
 					else
-					if(handlerAddress == (UINT32)MWA_RAMROM)
+					if(handlerAddress == (size_t)MWA_RAMROM)
 						strcpy(desc, "RAMROM");
 					else
 						strcpy(desc, "CUSTOM");
